Refactor reveal_crv3_commits()
