+++ pallets/subtensor/src/coinbase/mod.rs
@@ -1,5 +1,6 @@
 use super::*;
 pub mod block_emission;
 pub mod block_step;
+pub mod reveal_commits;
 pub mod root;
 pub mod run_coinbase;
